+++ worlds/pokemon_bw/generate/pokemon/species.py
@@ -10,8 +10,14 @@ def generate_species_data(world: "PokemonBWWorld") -> dict[str, SpeciesEntry]:
     from .evolutions import randomize_evolutions, replace_evolutions
     from .plando import plando_evolutions_override, plando_evolutions_append
     from .base_stats import randomize_stats_post_evo, randomize_stats_pre_evo
+    from .catch_rates import randomize_catch_rates
 
     all_species = {name: SpeciesEntry(name, data) for name, data in species.by_name.items()}
+
+    # Dependencies:
+    # Base stats - Evolutions
+    # Evolutions - Types, base stats
+    # Catch rates - Evolutions, base stats
 
     randomize_stats_pre_evo(world, all_species)
     replace_evolutions(world, all_species)
@@ -19,5 +25,6 @@ def generate_species_data(world: "PokemonBWWorld") -> dict[str, SpeciesEntry]:
     randomize_evolutions(world, all_species)
     plando_evolutions_append(world, all_species)
     randomize_stats_post_evo(world, all_species)
+    randomize_catch_rates(world, all_species)
 
     return all_species