+++ worlds/pokemon_bw/patch/procedures/write_stats.py
@@ -13,12 +13,14 @@ def write_patch(bw_patch_instance: "PokemonBWPatch", opened_zipfile: zipfile.Zip
     for species, data in bw_patch_instance.world.species_entries.items():
         if data.form and not data.is_custom_form:
             continue
-        byt = bytearray(7)
+        byt = bytearray(8)
         if data.write & 0b100:
             byt[:6] = (data.base_hp, data.base_attack, data.base_defense,
                        data.base_speed, data.base_sp_attack, data.base_sp_defense)
         if data.write & 0b1:
             byt[6] = data.evolution_stage
+        if data.write & 0b1000:
+            byt[7] = data.catch_rate
         if data.write & 0b101:
             opened_zipfile.writestr(f"stats/{max(data.dex_number, data.custom_form_file)}", bytes(byt))
 
@@ -35,7 +37,8 @@ def patch(rom: NintendoDSRom, world_package: str, bw_patch_instance: "PokemonBWP
 
             for j in range(6):
                 byt[j] = loaded[j] or byt[j]
-            byt[9] = byt[6]
+            byt[9] = byt[6] or byt[9]
+            byt[8] = byt[7] or byt[8]
 
             narc.files[i] = bytes(byt)
             files_dump[f"a016/{i}"] = narc.files[i]